backend/requirement.txt
Flask==2.0.2
pymongo==3.12.0
mongoengine==0.24.0
python-dotenv==0.19.2